Build: Make it possible to run MSAN in GN
MSAN has false positives due to b/234876100, but it works correctly for
most of the Pigweed codebase. This change makes it possible to use the
top-level `msan` target, but does not add msan to the default build.
Previously, the target was present, but was always a no-op.

Change-Id: I6811bb0f24f879f73c1bbf6afa3ffbf928f6d963
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/167112
Commit-Queue: Auto-Submit <auto-submit@pigweed-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Armando Montanez <amontanez@google.com>
Pigweed-Auto-Submit: Wyatt Hepler <hepler@google.com>