pw_cli: Narrow logic around colors
We would always use colors if PW_USE_COLOR was set to a "true" value, or
if it was not set, or if we're connected to a terminal. Change that
logic to default to using colors if connected to a terminal, and
overriding that default if PW_USE_COLOR is set.

Change-Id: I057c76c1b0c72f96c5d0177bc84ffb45ea6d57ec
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/173232
Commit-Queue: Auto-Submit <auto-submit@pigweed-service-accounts.iam.gserviceaccount.com>
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Ted Pudlik <tpudlik@google.com>
Pigweed-Auto-Submit: Rob Mohr <mohrr@google.com>