pw_emu: core: Let the OS terminate foreground emulator processes
When the CLI runs the emulator in foreground there is no need to
explicitly terminate the emulator processes, the OS will do that for
us when the CLI process is terminated as long as the CLI runs attached
to a console.

Foreground CLI tests do not have a console attached (due to missing
pty support on Windows) and instead are using plain pipes. Because of
that logic was added to the core code to terminate the emulator
processes when it was running in foreground.

However, foreground tests were refactored to run the CLI as a session
leader and that allows clean termination for emulator processes
directly from the tests.

Thus it is no longer necessary to explicitly terminate emulator
processes in the core code when running in foreground.

Change-Id: Ib16578a94a757dcde3bf3744253745a9de52ffad
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/175638
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Octavian Purdila <tavip@google.com>