pw_allocator: Refactor Block to use encoded offsets
This CL refactors `Block` into `Block<UintType, kMaxSize>`. Instead of
using next and previous block pointers, it now encodes offsets and flags
in fields of type `UintType`. Depending on `kMaxSize`, there may be bits
of `UintType` that are not needed to address blocks. These bits are
available to be used as custom flags.

Change-Id: Ic979f6d6982088de810bdd9b793e1e115aef65eb
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/176578
Reviewed-by: Taylor Cramer <cramertj@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>