pw_env_setup: Make pigweed_environment.gni content gni-relative
This is useful for Matter which tries to reuse a single environment
between multiple GN roots. The path is ambiguous if specified relative
to GN root, but is unambiguous if specified relative to the gni file.

The paths are turned back into "source absolute paths" via
get_path_info, so the variables actually have the same value as before
whenever the gni file is loaded from //build_overrides/.

However, if the gni file is imported indirectly from e.g.
//third_party/connectedhomeip/build_overrides, as the Matter sample
projects do, then the environment will be at
//third_party/connectedhomeip/.environment, which is the correct
location.

(Contrary to what it may seem, get_path_info(path, "abspath") outputs a
path relative to the GN --root, i.e. exactly what was in this file
before. So the change should be safe for pigweed & its other users).

Change-Id: I48f85d81ca1cedd05d0a2711420f5e4c9a1e7a37
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/180991
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Michael Spang <spang@google.com>