pw_bluetooth_sapphire: Move LegacyLowEnergyScanner impl to base class
This change moves much of the LegacyLowEnergyScanner implementation up
to its base class, LowEnergyScanner. This change also updates tests to
be generic type parameterized tests with LegacyLowEnergyScanner being
the only current tested implementation. This is in preparation of adding
a new ExtendedLowEnergyScanner subclass to LowEnergyScanner. There are
no logic changes in this diff. The only changes are refactoring related.

Bug: 305975969
Change-Id: I1599c1f1aaf866d0e20dbf48bed84f7ec0493154
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/182017
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Faraaz Sareshwala <fsareshwala@google.com>