pw_allocator: Add AllocationTestHarness
This CL adds a test utility class that can generate requests to
allocate, deallocate, and reallocate memory. The class has a
configurable number of allocated pointers it can record in a vector.
Finally, this class has methods to handle these requests that can be
used to programatically drive the allocator with a large number of
generated calls as part of a fuzz, perf, and/or stress test.

Change-Id: I861b57a21c8ef541da70058fc4bfd85a82cf494e
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/180532
Reviewed-by: Taylor Cramer <cramertj@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>