pw_allocator: Add extra bytes to Block
The CL adds a new template parameter, kNumExtraBytes, to `Block`, and
refactors the existing parameters. The new parameter can be used to
reserve space in the block's header that can be used to store and fetch
allocation metadata.

Additionally, the `kMaxSize` parameter has been dropped and replaced by
a parameter for `kNumFlags`. This allows better elision of flag-related
code when it is not needed, and encourages downstream projects to define
fewer block types rather than one for each allocation pool size.

Change-Id: I3530d94b1302e0dada9ad1f2f749ee94498ebb41
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/185954
Reviewed-by: Taylor Cramer <cramertj@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>