pw_bluetooth_sapphire: Advertise FCS Option
Not advertising this option causes bad behavior by some peer stacks,
they do not send FCS bits even though they should when this is the
default (16-bit FCS).

Add the FCS Option to the options parsed and understood, but do not
chaneg behavior.  Not sending the FCS option means the peers are
required to send us FCS as both sides must agree to turn it off and it
is non-negotiatiable.

Bug: b/318002648
Test: updated BrEdrDynamicChannelTest
Change-Id: I9be655c75771a36d0e670ee0e90e28e899997756
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/186850
Commit-Queue: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>