pw_protobuf: Fix out-of-range read
A failed `varint::Read` led to a nested decoder's position to not be
updated when the stream was (partially) advanced. This led to destructor
reading past the end of the stream when `CloseNestedDecoder` was
invoked.

To fix this, `varint::Read` now always returns the number of bytes read
and `StreamDecoder` always updates its position, even on error.

Bug: b/314803709
Change-Id: Id07ea172030ada376d5ce1ce3844e0014edbe4dc
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/188095
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>