pw_toolchain: Internal wrapper for casting between siblings
Casting between sibling types (types with a common base) can be useful
for efficiently exposing different interfaces for the same data.
However, it is invalid to access an object through a reference to an
incompatible type in C++. This change introduces a SiblingCast operator
that checks that the types are compatible and uses std::launder to avoid
undefined behavior.

Fixes: b/319144706
Change-Id: I69e48c662ef84dd747211b521d6951c10774c442
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/189014
Reviewed-by: Armando Montanez <amontanez@google.com>
Pigweed-Auto-Submit: Wyatt Hepler <hepler@google.com>
Commit-Queue: Wyatt Hepler <hepler@google.com>