pw_transfer: Allow setting a transfer resource size in C++
This adds interfaces to both the server-side `TransferHandler` and
client-side `TransferHandle` in the C++ transfer code which allow
users to specify the total size of their transfer resource.

When set, the size will be reported through the transfer protocol
using the `remaining_size` field in transmitted chunks.

Bug: 319731837
Change-Id: Icd0ae7d2547e7cb305e69d1e184485166dda2512
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/189095
Reviewed-by: Wyatt Hepler <hepler@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>
Reviewed-by: Jordan Brauer <jtbrauer@google.com>