pw_allocator: Remove total_bytes metric
Previously, `pw::allocator::Metrics` had a field for `total_bytes`. This
is not particularly interesting, as it is a value that isn't expected
to change. It also is a bit difficult, as a `TrackingAllocator` has no
way to interrogate an `Allocator` it is handed for its implementation-
specific capacity. As a result, this CL removes the field.

Change-Id: If8f587f2963d763cd01d48836ad111bf54634296
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/193251
Reviewed-by: Ewout van Bekkum <ewout@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>