pw_allocator: Make TrackingAllocator correct by construction
This CL removes the `Init` method from `TrackingAllocator` and
requires objects of that type to always be constructed with a valid
allocator. Previously, allowing the pointer to the wrapped allocator
to be null was believed to be needed to allow trackers around allocators
that might be optional depending on the platform. However, there is
already an allocator that can be used for this purpose: the
NullAllocator. As a result, the wrapped allocator can be a reference
instead of a pointer, and several conditional statement can be removed.

Change-Id: If0fd09348e7ca88c04c279e42c3ab94dfdab83da
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/193330
Reviewed-by: Ewout van Bekkum <ewout@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>