pw_stream_uart_mcuxpresso: Implement init / deinit
This change implements the UartStreamDmaMcuxpresso::Init
function to initialize the USART and DMA engine.

The UsartDmaTxData and UsartDmaRxData encapsulate the state
that is required to handle a TX or RX transaction.

A subsequent change will start the TX receive DMA transaction
from this function to read into the ring buffer.

Bug: 325514698
Change-Id: I9b136e16c0cffa358a26cfc675a27e5308090743
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/192891
Reviewed-by: Austin Foxley <afoxley@google.com>
Commit-Queue: Christoph Klee <chklee@google.com>
Reviewed-by: Jonathon Reinhart <jrreinhart@google.com>