pw_transfer: Fix Java client timeouts in terminating state
If a transfer fails to receive a COMPLETION_ACK during a terminating
state, it should finish with its original status. The Java client was
not properly handling this, instead failing with a DEADLINE_EXCEEDED
as in other timeout cases. This corrects the behavior of the client.

Change-Id: I7986ab262d828d9ca30caf4da421bf589f02913c
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/195123
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Erik Gilling <konkers@google.com>
Reviewed-by: Jordan Brauer <jtbrauer@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>