pw_allocator: Move code snippets from docs to examples
What is better in than code samples in module documentation? Example
code that is actually compiled and tested as part of the build and the
included into the documentation!

This CL creates standalone examples as tests for all the code that
appears in the pw_allocator docs. This helps ensures the remain fresh
and accurate.

Bug: b/328076428
Change-Id: I853a21801c56093b8448e04f84d92e5245eab116
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/195190
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Taylor Cramer <cramertj@google.com>