pw_allocator: Add Capabilities
This CL adds the ability for derived allocators to provide a set of
capabilities, to the base allocator class. This allows both the base
class and forwarding allocators to inquire about optional methods
implemented or optional behaviors desired.

Currently, this is only used to enforce that the primary allocator for a
`FallbackAllocator` can `Query`. In subsequent CLs, it will also be used
to determine how `TrackingAllocator` should store allocation details,
and allow `BumpAllocator` to skip destroying objects in `Delete`.

Change-Id: Ia1a7069fe839126e40eb943e599d84a9ad437567
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/197133
Reviewed-by: Taylor Cramer <cramertj@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>