nanopb: Fix nanopb_pb2.py generation
Since nanopb 0.4.7, importing nanopb/generator/proto/__init__.py is not
sufficient to generate nanopb_pb2.py.

The more robust way to generate this file across nanopb versions is to
import nanopb_generator.py. This approach works with nanopb versions
>=0.4.1, and the existing Pigweed build infrastructure does not work
with nanopb versions prior to 0.4.2.

Add an assert to check that the file is actually generated so that we
can catch this error in the future if a future version of nanopb breaks
this script.

Test: Ran generate_nanopb_proto.py in a project with various versions of
  nanopb (0.4.1, 0.4.2, 0.4.6, 0.4.8) and confirmed that nanopb_pb2. was
  generated.
Change-Id: Ia39c60ae202fc1f0ff5da4221e983fcdf20e3c5e
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/200772
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Armando Montanez <amontanez@google.com>
Commit-Queue: Kesavan Yogeswaran <hikes@google.com>