pw_bluetooth: Store length max in virtual field
The LE Set Extended Advertising Data Command and LE Set Extended
Advertising Data Command have data length fields with a maximum
value. We had placed this maximum value restriction in Emboss via a
`[requires]` block. However, I'd like to be able to access the maximum
value of the restriction programmatically without having to redefine
it. This change places the maximum value literal in a virtual field so
that it is accessible programmatically. Due to some Emboss limitations,
we also have to rework how we apply the `[requires]` block.

Test: hardlink pigweed file into Fuchsia's third_party repository, build
Change-Id: I72fd5d9b8ebbfc292d70e8ddbf7c24c63c08a066
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/201794
Commit-Queue: Faraaz Sareshwala <fsareshwala@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Reviewed-by: Lulu Wang <luluwang@google.com>