pw_transfer: Implement adaptive windowing in C++
This implements adaptive transfer windowing in the C++ transfer client
and server. Window size is changed to be a multiple of the max chunk
size, and is increased in response to successful receipt of data and
decreased with packet loss. This follows an algorithm similar to TCP's
congestion control.

Change-Id: Ifba5cceaa79c83f6d44ebb2b12359f3ef0b1a263
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/146392
Commit-Queue: Alexei Frolov <frolv@google.com>
Reviewed-by: Keir Mierle <keir@google.com>