pw_transfer: Implement adaptive windowing in Python
This implements adaptive transfer windowing in the Python transfer
client. Window size is changed to be a multiple of the max chunk size,
and is increased in response to successful receipt of data and decreased
with packet loss. This follows an algorithm similar to TCP's
congestion control.

Change-Id: Ibeec9268abc33d7a2114df0c8ab67ce0f92d5920
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/147510
Reviewed-by: Taylor Cramer <cramertj@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>