pw_transfer: Implement adaptive windowing in Java
This implements adaptive transfer windowing in the Java transfer client.
Window size is changed to be a multiple of the max chunk size, and is
increased in response to successful receipt of data and decreased with
packet loss. This follows an algorithm similar to TCP's congestion
control.

Change-Id: I666a1acb2e78897718f2b729069a26a218fae6a7
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/147511
Commit-Queue: Alexei Frolov <frolv@google.com>
Reviewed-by: Taylor Cramer <cramertj@google.com>