pw_allocator: Move block allocators to separate files
This CL separates out the individual block allocators into their own
files and build targets, ahead of the addition of a
BucketBlockAllocator.

The most substantial part of this change was the reworking of the unit
tests to be more explicitly declared instead of relying on macros for
code generation.

Change-Id: I4f200eb3d3282de9bd3e367d4a0f95610e913a4b
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/198154
Reviewed-by: Taylor Cramer <cramertj@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>