pw_transfer: Attempt to recover when receiving invalid size
This updates the C++ transfer code to not immediately terminate a
receive transfer if the transmitter send more data than was requested.
Instead, it will resend transfer parameters and attempt to recover.

Previously, this state was assumed to indicate a bug with the
transmitter, but with the addition of adaptive windowing, there are
legitimate cases where a window may shrink in size with in-flight
packets from the previous window.

To avoid an infinite retry loop if the transmitter is actually
functioning improperly, recovery attempts are limited to the transfer's
lifetime retry count.

Change-Id: I517b40fdaf8ce49075496e4f1b3a04714e974a06
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/206632
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jordan Brauer <jtbrauer@google.com>
Commit-Queue: Alexei Frolov <frolv@google.com>