pw_i2c_mcuxpresso: Fix Clang compilation
This commit fixes compilation on Clang, by moving the PW_GUARDED_BY
macro to be in the correct location with respect to a field with a
default value.

When compiling with Clang, the PW_GUARDED_BY macro is not a no-op, so
the placement of the attribute it expands to is important.

Tested: Compiles with Clang
Change-Id: I25e7d792af942b3073d183aef743f7444a0d14b2
Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/209191
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Ted Pudlik <tpudlik@google.com>
Commit-Queue: Auto-Submit <auto-submit@pigweed-service-accounts.iam.gserviceaccount.com>
Pigweed-Auto-Submit: Eli Lipsitz <elipsitz@google.com>